Move `check_for_conflict_markers` to `status.py` for shared use
